Update AbstractRPMMojo.java
add the holder for the rpmBuildExe property, and set the default to 'rpmbuild' so that existing code will not break.  allows the user to set a property for their machine to provide the full path to the executable.  eg: in case they have multiple versions of rpmbuild, or, if /bin/sh can not for some reason, find it on the search path.